build: turn off default features in git2 (#407)
`ssh` and `https` are `default` features of `git2`:
<https://docs.rs/crate/git2/latest/features#default>.

Building `headson` can fail if Perl isn't available, as openssl is
built, which needs Perl. `headson` doesn't need git network access, so
it seems straightforward and safer to just remove the `ssh` and `https`
network features.

I left the `vendored-` libs in as they're still required for successful
independent compilation it seems.